+++ tests/Discussion/DiscussionServiceTest.php
@@ -104,7 +104,7 @@ final class DiscussionServiceTest extends ApiTestCase
     {
         $this->expectException(CompanyHasNoAdministrator::class);
 
-        $this->buildDiscussionService()->startDiscussionWithVendor(1);
+        $this->buildDiscussionService('customer-1@world-company.com', 'password-customer-1')->startDiscussionWithVendor(1);
     }
 
     public function testStartDiscussionOnInexistentVendor()
@@ -429,6 +429,87 @@ MSG;
         /** @var Response $response */
         $response = static::$historyContainer[2]['response'];
         $this->assertSame(201, $response->getStatusCode());
+    }
+
+    public function testStartDiscussionWithCustomer(): void
+    {
+        $discussionService = $this->buildDiscussionService('vendor@world-company.com', 'Windows.98');
+
+        $discussion = $discussionService->startDiscussionWithCustomer(3);
+
+        $expectedDiscussion = new Discussion(
+            [
+                'id' => 2,
+                'recipient' => 'Paul Martin',
+                'productId' => 0,
+                'title' => 'Contact Paul Martin',
+                'unreadCount' => 0,
+            ]
+        );
+
+        static::assertSame($expectedDiscussion->getId(), $discussion->getId());
+        static::assertSame($expectedDiscussion->getRecipient(), $discussion->getRecipient());
+        static::assertSame($expectedDiscussion->getProductId(), $discussion->getProductId());
+        static::assertSame($expectedDiscussion->getTitle(), $discussion->getTitle());
+        static::assertSame($expectedDiscussion->getUnreadCount(), $discussion->getUnreadCount());
+
+        // Check that the user can access the discussion
+        $discussionService = $this->buildDiscussionService('user@wizaplace.com', 'Windows.98');
+        $discussions = $discussionService->getDiscussions();
+
+        static::assertContainsOnly(Discussion::class, $discussions);
+        $discussion = \end($discussions);
+        static::assertSame($expectedDiscussion->getId(), $discussion->getId());
+    }
+
+    public function testStartDiscussionFromOrderWithCustomer(): void
+    {
+        $this->discussionService = $this->buildDiscussionService('vendor@world-company.com', 'Windows.98');
+
+        $discussion = $this->discussionService->startDiscussionOnOrderWithCustomer(2, 3);
+
+        $expectedDiscussion = new Discussion(
+            [
+                'id' => 2,
+                'recipient' => 'Paul Martin',
+                'productId' => 0,
+                'orderId' => 2,
+                'title' => 'A propos de la commande 2',
+                'unreadCount' => 0,
+            ]
+        );
+
+        static::assertSame($expectedDiscussion->getId(), $discussion->getId());
+        static::assertSame($expectedDiscussion->getRecipient(), $discussion->getRecipient());
+        static::assertSame($expectedDiscussion->getProductId(), $discussion->getProductId());
+        static::assertSame($expectedDiscussion->getOrderId(), $discussion->getOrderId());
+        static::assertSame($expectedDiscussion->getTitle(), $discussion->getTitle());
+        static::assertSame($expectedDiscussion->getUnreadCount(), $discussion->getUnreadCount());
+    }
+
+    public function testStartDiscussionFromOrderWithCompany(): void
+    {
+        $this->discussionService = $this->buildDiscussionService('user@wizaplace.com', 'Windows.98');
+
+        $discussion = $this->discussionService->startDiscussionOnOrderWithCompany(1, 3);
+
+        $expectedDiscussion = new Discussion(
+            [
+                'id' => 9,
+                'recipient' => 'The World Company Inc.',
+                'productId' => 0,
+                'orderId' => 1,
+                'title' => 'A propos de la commande 1',
+                'unreadCount' => 0,
+            ]
+        );
+
+        static::assertSame($expectedDiscussion->getId(), $discussion->getId());
+        static::assertSame($expectedDiscussion->getRecipient(), $discussion->getRecipient());
+        static::assertSame($expectedDiscussion->getProductId(), $discussion->getProductId());
+        static::assertSame($expectedDiscussion->getOrderId(), $discussion->getOrderId());
+        static::assertSame($expectedDiscussion->getTitle(), $discussion->getTitle());
+        static::assertSame($expectedDiscussion->getUnreadCount(), $discussion->getUnreadCount());
     }
 
     private function buildDiscussionService($email = 'customer-1@world-company.com', $password = 'password-customer-1'): DiscussionService