+++ CHANGELOG.md
@@ -1,6 +1,40 @@
 
 # Changelog
 
+## 1.146.0
+
+### New features
+
+- Added optional parameter `$companyIds` to `Wizaplace\SDK\Pim\Product\ProductService::updateStock`
+- In the `Wizaplace\SDK\Discussion\Discussion` class, added the "orderId" field which is returned by the API endpoint.
+- Added `Wizaplace\SDK\Discussion\DiscussionService::startDiscussionOnOrderWithCustomer`
+- Added `Wizaplace\SDK\Discussion\DiscussionService::startDiscussionOnOrderWithCompany`
+- Fixed some @param comments which were plurals
+- Added a default _Accept_ header set to _application/json_.
+
+## 1.144.0
+
+### New features
+
+- Added `\Wizaplace\SDK\Company\AbstractCompanyRegistration::getInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\AbstractCompanyRegistration::setInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\Company::getInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\Company::setInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\CompanyUpdateCommand::getInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\CompanyUpdateCommand::setInvoicingDisabled`
+
+## 1.143.0
+
+### New features
+
+- Added `Wizaplace\SDK\Discussion\DiscussionService::startDiscussionWithCustomer`
+
+## 1.142.0
+
+### New features
+
+- In the `Wizaplace\SDK\User\User` class, added the "passwordExpiryTimeLeft" field which is returned by the API endpoint.
+
 ## 1.139.0
 - Added an interface to enable event dispatching before and after each request to the API.